+++ satella/configuration/schema/from_json.py
@@ -46,6 +46,8 @@ def _get_descriptor_for_dict(key: str, value: dict) -> Descriptor:
                 args = eval('lambda x: ' + value['expr'], dict_),
             else:
                 args = y,
+        elif type_ == 'file_contents':
+            args = value.get('encoding', None),
         elif type_ == 'union':
             args = [_get_descriptor_for('', x) for x in value.get('of', [])]
         optional, default = False, None