+++ satella/configuration/sources/__init__.py
@@ -4,12 +4,14 @@ from . import envvars
 from . import file
 from . import format
 from . import from_dict
+from . import object_from
 from .base import *
 from .derivative import *
 from .envvars import *
 from .file import *
 from .format import *
 from .from_dict import *
+from .object_from import *
 
 __all__ = format.__all__ + envvars.__all__ + derivative.__all__ + file.__all__ + base.__all__ + \
-          from_dict.__all__
+          from_dict.__all__ + object_from.__all__