+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.16.2
 
 * added optional argument to `File.get_value`
+* added `file_contents` schema 
+* added nested configuration sources